+++ docs/change-log.md
@@ -5,6 +5,7 @@
 ### Added
 
 -   Header module: support for sub-actions.
+-   Form module: chips field.
 
 ## [1.0.0] - 2026-07-20